Refactor: Coordinate 타입 변경

+++ favoriteToilet/favoriteToilet/Source/Model/Domain/Coordinate.swift
@@ -8,6 +8,6 @@
 import Foundation
 
 struct Coordinate {
-    let latitude: Float
-    let logitude: Float
+    let latitude: Double
+    let logitude: Double
 }
